requirements.txt
# requirements.txt

#pip install -r requirements.txt
Django
psycopg2
selenium
djangorestframework
requests
pytest-django
pytest-asyncio


